pyproject.toml
[project]
name = "moblog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "cryptography>=45.0.3",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "orjson>=3.10.18",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "tortoise-orm[aiomysql]>=0.25.0",
    "uvicorn>=0.34.2",
    "uvloop>=0.21.0",
    "xxhash>=3.5.0",
]

[dependency-groups]
dev = ["isort>=6.0.1"]
